% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_examples.R
\name{create_example_1}
\alias{create_example_1}
\title{Copy files for example-1 to directory}
\usage{
create_example_1(path = ".")
}
\arguments{
\item{path}{An optional character string, holding the relative path to
to the directory, where the example folder should be copied to.
The default value is the current working directory.}
}
\description{
Copy files for example-1 to directory
}
