% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringan_inf_mr.R
\name{daemon_stop}
\alias{daemon_stop}
\title{Stop rendering with \code{\link[=inf_mr]{inf_mr()}}}
\usage{
daemon_stop(which = daemon_list())
}
\arguments{
\item{which}{A integer vector of the server IDs; by default, IDs of all
existing servers in the current R session obtained from
\code{daemon_list()}, i.e., all daemon servers will be stopped by default.}
}
\description{
This function was taken from the \link[servr:daemon_stop]{servr package}
(see \code{\link[servr:daemon_stop]{servr::daemon_stop()}} for more details).

This function stops the background process (\strong{http server} and rerendering)
started by calling \code{\link[=inf_mr]{inf_mr()}}.
}
