% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringan_inf_mr.R
\name{inf_mr}
\alias{inf_mr}
\title{Render and serve paginated markdown files}
\usage{
inf_mr(moon, cast_from = ".", ...)
}
\arguments{
\item{moon}{The input Rmd file path (if missing and in RStudio, the current
active document is used).}

\item{cast_from}{The root directory of the server.}

\item{...}{Passed to \code{rmarkdown::\link[rmarkdown]{render}()}.}
}
\description{
This function was taken from the \link[xaringan:inf_mr]{xaringan package}
(see \code{\link[xaringan:inf_mr]{xaringan::inf_mr()}} for more details).

This function starts a background process, which does the following things:
\itemize{
\item if an \strong{Rmd} file is changed, automatically rerender the \strong{html} file
\item serve the \strong{html} file and all included files (images, stylesheets, javascript libraries etc.)
via an \strong{http server}.
}

The rendered \strong{html} file can be viewed by opening the \strong{url} shown in the
console output (e.g. \verb{127.0.0.1:4321/my_dir/my_report.html}) in \strong{Google chrome}.

In order to stop the background process, please call \code{\link[=daemon_stop]{daemon_stop()}}.

Be aware that, if the \strong{Rmd} files are not located in the working directory,
then you have also pass in \code{knit_root_dir}
(e.g.: \code{inf_mr("my_dir/my_report.Rmd", knit_root_dir = normalizePath("my_dir"))}).
\strong{current working directory}.
}
