% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdmux.R
\name{insert_fig_footer}
\alias{insert_fig_footer}
\title{Allow footer text for figures to be set with \code{fig.footer} chunk option.}
\usage{
insert_fig_footer(x, options)
}
\description{
Allow footer text for figures to be set with \code{fig.footer} chunk option.
}
