% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdmux.R
\name{html_rmdmux_paged}
\alias{html_rmdmux_paged}
\alias{html_rmdmux_plain}
\title{HTML output formats for reports}
\usage{
html_rmdmux_paged(
  ...,
  template = pkg_resource("html", "rmdmux_paged.html"),
  css = NULL,
  toc = TRUE,
  md_extensions = "-autolink_bare_uris",
  number_sections = TRUE,
  fig_caption = TRUE
)

html_rmdmux_plain(
  ...,
  template = pkg_resource("html", "rmdmux_plain.html"),
  css = NULL,
  toc = TRUE,
  md_extensions = "-autolink_bare_uris",
  number_sections = TRUE,
  fig_caption = TRUE
)
}
\arguments{
\item{...}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link[rmarkdown]{html_document_base}}}

\item{template}{Pandoc template to use for rendering. Pass "default" to use
the rmarkdown package default template; pass \code{NULL} to use pandoc's
built-in template; pass a path to use a custom template that you've created.
Note that if you don't use the "default" template then some features of
\code{html_document} won't be available (see the Templates section below for
more details).}

\item{css}{CSS and/or Sass files to include. Files with an extension of .sass
or .scss are compiled to CSS via \code{sass::sass()}. Also, if \code{theme} is a
\code{\link[bslib:bs_theme]{bslib::bs_theme()}} object, Sass code may reference the relevant Bootstrap
Sass variables, functions, mixins, etc.}

\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition of R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for
additional details.}

\item{number_sections}{\code{TRUE} to number section headings}

\item{fig_caption}{\code{TRUE} to render figures with captions}
}
\description{
The following \strong{html output formats} are included in the \strong{rmdmux} package:
\itemize{
\item \code{\link[=html_rmdmux_paged]{html_rmdmux_paged()}}: A \strong{paginated html} output, which can be printed to \strong{pdf}
\item \code{\link[=html_rmdmux_plain]{html_rmdmux_plain()}}: A \strong{plain html} output, which is fully accessible
(eg. for screen readers etc.)
All output formats of the \strong{rmdmux} package support the usage of
\link[styledTables:StyledTable-styled_table-method]{styledTables}.
For a better understanding, see section \strong{Supported yaml parameters} or
have a look at the  look at the minimal example, which can be created with the command
\code{\link[=create_example_1]{create_example_1()}}.
}
}
\section{Supported yaml parameters}{

The following yaml parameters are supported by the \code{rmdmux} output formats:
\itemize{
\item \code{title}: Title of the document
\item \code{subtitle}: Optional subtitle of the document
\item \code{author}: Optional field holding one ore more author names.
Either a character vector or an unnamed list, where each list entry has
the entries \code{name} and \code{affiliation}.
\item \code{date}: Optional date string
\item \code{lang}: HTML language definition. Inserted in \verb{<html lang="XX">} tag of the html document.
\item \code{params}: Optional list of parameters, which should be available inside of the \strong{rmd chunks}.
\item \code{output}: Specify one or more output formats:
\itemize{
\item \code{html_rmdmux_paged} or \code{html_rmdmux_plain}:
\itemize{
\item \code{css}: Additional \strong{css}, \strong{scss} or \strong{sass} files, which should
be appended to the output file.
\item \code{toc}: Should a \strong{table of contents} be included?
\item \code{md_extensions: -autolink_bare_uris}: Prohibit pandoc from automatically
replacing all urls in the \strong{yaml} header by \verb{<a href="...">} elements
\item \verb{pandoc_args: ["-M", "paged-footnotes=true"]}: This option shoud only be
used for \code{html_rmdmux_paged}: This causes the footnotes to be printed
at the bottom of each page and not as a collection at
the end of the document.
}
}
\item \code{lot}: Should a \strong{list of table} be included?
\item \code{lof}: Should a \strong{list of figures} be included?
\item \code{toc-title}: Title of the \strong{table of contents}
\item \code{lot-title}: Title of the \strong{list of tables}
\item \code{lof-title}: Title of the \strong{list of figures}
\item \code{paged-footnotes}: Print footnotes at the bottom of each page
(Cauton: This option is valid only for \code{html_rmdmux_paged} =>
Better use \code{pandoc_args} solution listed above)
\item \code{title-page}: Optional list object defining the title page
\itemize{
\item \code{cover-image}: Optional list object defining a cover image, which will
be positioned below the report title.
\itemize{
\item \code{src}: file path of the cover image
\item \code{alt}: alternative text for the cover image
}
\item \code{cover-logo}: Optional list object defining a cover logo, which will
be positioned in the page corner of the title page of the paginated output:
\itemize{
\item \code{src}: file path of the cover image
\item \code{alt}: alternative text for the cover image
}
\item \code{footer}: Optional list object defining the footer of the title page
\itemize{
\item \code{left}: Optional text for the left part of page footer
\item \code{middle}: Optional text for the center part of page footer
\item \code{right}: Optional text for the right part of the page footer
}
}
\item \code{imprint}: Configure the \strong{imprint} section
\itemize{
\item \code{heading}: Optional string holding the title of the \strong{imprint} section
\item \code{title-label}: Optional label for the title field
\item \code{edition-label}: Optional label for the edition field
\item \code{edition}: Optional string holding information about the report edition
\item \code{publisher-label}: Optional label for the publisher field
\item \code{publisher}: Optional character vector holding information about the publisher
\item \code{address-label}: Optional label for the address field
\item \code{address}: Optional character vector holding the address as multiple lines of text
\item \code{author-label}: Optional label for the author field
\item \code{editor-label}: Optional label for the editor field
\item \code{editor}: Optional field holding one ore more editor names.
Either a character vector or an unnamed list, where each list entry has
the entries \code{name} and \code{affiliation}.
\item \code{email-label}: Optional label for the email field
\item \code{email}: Optional string holding the email address
\item \code{phone-label}: Optional label for the phone number field
\item \code{phone}: Optional string holding the phone number
\item \code{url-label}: Optional label for the \strong{url} field
\item \code{url}: Optional string holding the \strong{url} of the publisher
\item \code{url-qrcode}: Optional list object defining a QR-code image, which will
be positioned next to the \code{url}.
\itemize{
\item \code{src}: file path to the image file for the QR-code
\item \code{alt}: alternative text for the QR-code image
}
\item \code{date-label}: Optional label for the date field
\item \code{place-label}: Optional label for the place field
\item \code{place}: Optional character vector holding information about the publishing place
\item \code{remark}: Optional character vector holding one ore more paragraphs
of remarks, which will be appended at theend of the imprint section
\item \code{copyright-label}: Optional label for the copyrights field
\item \code{copyright}: Optional character vector holding information about
credits and copyrights
\item \code{remarks}: Optional character vector holding additional remarks,
which will be included at the end of the imprint.
}
}
}

\section{Translation of report element labels}{

In order to translate report element labels into other languates
(e.g. \code{"Abbildung"} is german for \code{"Figure"} etc.) we use the \verb{_bookdown.yml}
file, in which we can configure all important report labels. This file
must be located in the same folder as the \strong{rmd} files.
For a detailed description see \url{https://bookdown.org/yihui/bookdown/internationalization.html}
}

